Added Element monster system on MonsterSetBase.xml, added WZTXTGeneral Loader from Class Definition, fix magicattack packet for S9

+++ MuEmu/Constants.cs
@@ -513,7 +513,7 @@ namespace MuEmu
         DevilSquare = 4,
     }
 
-    public enum DamageType : byte
+    public enum DamageType : ushort
     {
         Regular,
         Excellent = 2,
